apt-audit phase 2: adversarial verify round -- 45 refuted claims fixed
Five dump-and-probe verifiers re-derived 333 claims from the binaries (no
author rationale shared). 46 refutations, 45 mechanical fixes applied + the
CloneItem chain retype:

- SetDirtyState was SPLIT-BRAIN: it wrote the console dirty position (bit 25)
  while GetDirtyState/tick read the x64 bit 6 -- a set was invisible to every
  reader, and tick's opening gate also still read the console mask (no clip
  would ever tick). Both now on x64 bit 6 (XB1 tick sub_14085D620 `& 0x40`).
- CreateCharacterInst: the morph (type 8) inst was allocated base-sized (its
  +0x20 member landed OUTSIDE the block -- pool corruption); and unknown types
  fabricated a base inst where BOTH binaries return null. Fixed per the two
  factory bodies (X360 @0x82AFFF70, x64 sub_140835410).
- AptCharacterInst::CloneItem retyped to the x64 mangling (void, const-CIH/CIH
  pointer pair) and the pointer carried through AptCloneManagedItem end-to-end
  (the int middle leg truncated the dest pointer on x64).
- SetDepth/SetClipDepth + sibling setters aligned to their void x64 manglings.
- ~40 further offset/bit/signedness/side-effect refutations across the render,
  interpreter, value/GC and boot clusters (each fix carries its deciding asm).

Remaining SUSPECT findings (documented, deliberately not rushed): the _gotoAndX
boxed-string null guard, the Sprite/CustomControl setter signature parity, the
stackPop/SetCurItem void returns, gValuesToRelease as a pool-allocated
AptValueVector* (vs the static-instance model), and folding the two input-
recorder sinks into their real gAptFuncs slots (+0x18/+0x1C).

Gates: cl /c green on all touched TUs + header siblings; faithfulness PASS.

+++ src/SDKs/EATech/Apt/AptUpdate.cpp
@@ -202,7 +202,7 @@ static int AptUpdateRunTargetFrames(int nElapsedMs, int nDepthLayerMask, int nMa
         // single-threaded, so a plain add reproduces the observable state.
         // (The X360 also traps on nMsPerFrame == 0 -- twllei -- before the divide.)
         if (nMsPerFrame != 0
-            && (gnCurrUpdateTick - gnCurrRenderTickConsumed) / static_cast<int>(nMsPerFrame)
+            && static_cast<int>(static_cast<unsigned int>(gnCurrUpdateTick - gnCurrRenderTickConsumed) / nMsPerFrame)
                    < nMaxBankedFrames)
         {
             gnCurrUpdateTick += nMsPerFrame;
